datastore-api/requirements.txt
Jinja2 == 3.0.1
fastapi == 0.65.2
filelock == 3.0.12
h5py == 3.3.0
motor == 2.4.0
pydantic == 1.8.2
python-dotenv == 0.18.0
python-multipart == 0.0.5
pyvespa == 0.7.0
requests == 2.25.1
torch == 1.9.0
transformers == 4.8.2
uvicorn[standard] == 0.14.0
ElkJsonFormatter.tar.gz
